Use 'unsigned' rather than 'ushort' for SubMesh indices

+++ Components/MeshLodGenerator/src/OgreLodInputProviderBuffer.cpp
@@ -51,8 +51,8 @@ namespace Ogre
         size_t vertexCount = 0;
         size_t vertexLookupSize = 0;
         size_t sharedVertexLookupSize = 0;
-        unsigned short submeshCount = mBuffer.submesh.size();
-        for (unsigned short i = 0; i < submeshCount; i++)
+        unsigned submeshCount = mBuffer.submesh.size();
+        for (unsigned i = 0; i < submeshCount; i++)
         {
             const LodInputBuffer::Submesh& submesh = mBuffer.submesh[i];
             trianglesCount += getTriangleCount(submesh.operationType, submesh.indexBuffer.indexCount);
@@ -85,8 +85,8 @@ namespace Ogre
         data->mMeshName = mBuffer.meshName;
 #endif
         data->mMeshBoundingSphereRadius = mBuffer.boundingSphereRadius;
-        unsigned short submeshCount = mBuffer.submesh.size();
-        for (unsigned short i = 0; i < submeshCount; ++i)
+        unsigned submeshCount = mBuffer.submesh.size();
+        for (unsigned i = 0; i < submeshCount; ++i)
         {
             LodInputBuffer::Submesh& submesh = mBuffer.submesh[i];
             LodVertexBuffer& vertexBuffer =
